+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 - fix(components/table): prizmThGroup throws an error together with structure directives #510
 - fix(doc/select): set default value in example with transformer
 
+### For testers
+
+- !!! need check all overlay elements (sidebars, dialogs, confirms and so on) and overlay in tools. We changed core logic
+
 ## [2.1.2](https://github.com/zyfra/Prizm) (24-07-2023)
 
 ### Features